修复代码生成获取子包类，不让生成子包类

+++ src/main/java/com/xxbb/springbootapi/service/impl/CodeService.java
@@ -1,6 +1,7 @@
 package com.xxbb.springbootapi.service.impl;
 
 import com.xxbb.springbootapi.entity.GenRecord;
+import com.xxbb.springbootapi.interceptor.LegalException;
 import com.xxbb.springbootapi.service.ICodeGenerate;
 import com.xxbb.springbootapi.utils.code.CodeGen;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,11 +15,15 @@ import java.util.List;
  */
 @Service
 public class CodeService implements ICodeGenerate {
+    private static final String BASE_CLASS = "User,Authority,Role,Base,GenRecord,SecurityUserDetail,Api,AuthApi,Code,File,Common,WebSocket";
     @Autowired
     private GenRecordService genRecordService;
 
     @Override
     public Boolean generate(String entityName, Boolean isCover) {
+        if (BASE_CLASS.contains(entityName)) {
+            throw new LegalException("内置基础实体类，不允许该操作");
+        }
         genRecordService.add(new GenRecord().setEntityName(entityName));
         return CodeGen.build(entityName, isCover);//代码生成，参数：要生成代码的实体类名
     }
@@ -26,6 +31,9 @@ public class CodeService implements ICodeGenerate {
     @Override
     public Boolean delete(Integer id) {
         GenRecord genRecord = genRecordService.find(id);
+        if (BASE_CLASS.contains(genRecord.getEntityName())) {
+            throw new LegalException("内置基础实体类，不允许该操作");
+        }
         genRecordService.delete(genRecord.getId());
         return CodeGen.delete(genRecord.getEntityName());//代码删除，参数：要删除代码的实体类名
     }
